Added draft of redis buffer
